requirements.txt
discord.py
python-dotenv
aiohttp
Pillow
pytz
beautifulsoup4
requests